Plan colors only at high alignment
Color mixing of threads doesn't work well, in particular for low density
regions. For instance, green + red pretty much never results in a yellow
effect, I guess.

So we'll now only consider a 'primary' color when it is quite
aligned with the actual one (as per angle in Lab) and plan only one
primary color accordingly.