Improve coloring, magic constants
A little too many but it looks decent with those. The largest problem is
visualization (or planning..): when both a gray and a color is planned
then the debug view renders one on top of the other such that they do
not mix as separate threads would.

We could avoid having different colors on the same nodes?